Close the audio channel after no speech for 30 seconds.
This change also includes some refactoring that makes it
possible for HMM and Unit Selection engines to be loaded
at the same time, however no unit selection voices are
installed so that code won't have any effect. The goal
is to enable this in M31.

BUG=chromium:255669
TEST=Chromebook goes to sleep even when spoken feedback is on.

Change-Id: Ie0499a3c9c8cbce6ec12463a6d1778768a681224
Reviewed-on: https://chromium-review.googlesource.com/65291
Reviewed-by: David Tseng <dtseng@chromium.org>
Tested-by: David Tseng <dtseng@chromium.org>
Commit-Queue: Dominic Mazzoni <dmazzoni@chromium.org>