tree_status: fix handling of extended tree states
The current code doesn't handled throttled or maintenance tree states
correctly.  It tries to decode them as "neither" or "unknown".

Add a single dict to match correct statuses to the right color (and
update those colors to match the css used on the status servers),
rename "neither" to "throttled", and add a maintenance image.

BUG=None
TEST=ran against local app instance with different modes

Change-Id: If6ba428cb7291fc3b01ba2f83c77e43adbee5547
Reviewed-on: https://chromium-review.googlesource.com/176382
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>