tree_status: integrate sheriff support
The current icon in the browser does not have great visibility when
sheriffing and you're looking at other windows.  Leverage the Chrome
notification features to display a pop up whenever the tree changes
state.  That notification can also be clicked to be taken straight
to the status page.

We currently load the west coast & non-west coast sheriff rotations.
We do not look at the gardeners as they are not the ones responsible
for keeping the tree open.

BUG=None
TEST=reviewed icon behavior w/out username set -- had a notice & title text
TEST=reviewed icon behavior w/username set to a sheriff -- got notifications
TEST=reviewed icon behavior w/username set not to a sheriff -- got no notifications

Change-Id: Ifdfeee7ea9fef9ddfc63cf39cef501531fdb88da
Reviewed-on: https://chromium-review.googlesource.com/229640
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>