iter_viewer: update to handle Week/Phase
Now that the Iteration is no more, fetch the Week/Phase instead and
display that.  Other than mechanical shuffling, the largest change
here is the icon.  Since it's 4 digits now, we have to be a bit more
creative in how we display it in order to fit in the 19x19 square.
We do a couple of things:
 - split the number in two and stack them on top of each other
 - use a light background (that fads into the URL space)
 - use black text that runs to the edge of the icon (and re-uses the
   URL space for contrast)
 - use a larger font

BUG=None
TEST=loaded the icon and looked at the data

Change-Id: If585873c7b083b9f00e2223bc926515e651b2135
Reviewed-on: https://chromium-review.googlesource.com/256280
Reviewed-by: Alex Deymo <deymo@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>