boot_messages: Remove the recovery instruction in warn_dev message.
Currently on Freon boards, we don't have full text terminal support so the
instruction of "entering recovery mode by space" won't really work.

To avoid confusion, it seems better to remove the instruction for now. We can
add it back when the full support of space and Ctrl-D is ready.

BUG=chromium:479964
TEST=None

Change-Id: I25fa1980fcdc3bdf11dfb095aa219d21b5036bd6
Reviewed-on: https://chromium-review.googlesource.com/268800
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: David Sodman <dsodman@chromium.org>
Reviewed-by: Dharani Govindan <dharani@chromium.org>
Commit-Queue: Hung-Te Lin <hungte@chromium.org>