New Winky icc profile
This icc profile was provided by Samsung.  The name combines the
manufacturer id and product code of the winky display panel, so that the
color profile will only be applied to this panel.  This CL goes with
https://chromium-review.googlesource.com/#/c/272222/
and replaces
https://chromium-review.googlesource.com/#/c/272218/
which is out of date now that Samsung has provided a newer icc file and
we've updated our naming scheme.

BUG=chromium:468191
TEST=compile, run on winky: see 06af105c.icc in /usr/share/color/icc/

Change-Id: Ib30dfd60e8a46ed873e0e43283851fb991b1aec0
Reviewed-on: https://chromium-review.googlesource.com/282751
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>
Tested-by: Greg Levin <glevin@chromium.org>
Commit-Queue: Greg Levin <glevin@chromium.org>